Revert "XW-4257 | FV viewability AB test"
