+++ app/components/application-wrapper.js
@@ -1,13 +1,10 @@
 import {inject as service} from '@ember/service';
-import {reads, bool, equal, and, readOnly} from '@ember/object/computed';
+import {reads, bool, equal, and} from '@ember/object/computed';
 import Component from '@ember/component';
-import {computed, observer} from '@ember/object';
-import {scheduleOnce} from '@ember/runloop';
+import {computed} from '@ember/object';
 import $ from 'jquery';
 import {isHashLink} from '../utils/article-link';
 import {trackPerf} from '../utils/track-perf';
-import {inGroup} from '../modules/abtest';
-import {updateFeaturedVideoPosition} from '../modules/abtest/featured-video-render-order-helper';
 
 /**
  * HTMLMouseEvent
@@ -40,18 +37,14 @@ export default Component.extend({
 		'bfaaTemplate'
 	],
 	scrollLocation: null,
+	smartBannerVisible: false,
 	firstRender: true,
 
 	ads: service(),
 	currentUser: service(),
-	smartBanner: service(),
 	fastboot: service(),
 	logger: service(),
 	wikiVariables: service(),
-
-	smartBannerVisible: readOnly('smartBanner.smartBannerVisible'),
-	shouldShowFandomAppSmartBanner: readOnly('smartBanner.shouldShowFandomAppSmartBanner'),
-	isFandomAppSmartBannerVisible: readOnly('smartBanner.isFandomAppSmartBannerVisible'),
 
 	dir: reads('wikiVariables.language.contentDir'),
 
@@ -67,13 +60,12 @@ export default Component.extend({
 		return `${vertical}-vertical`;
 	}),
 
-	fandomAppSmartBannerObserver: observer('isFandomAppSmartBannerVisible', () => {
-		if (inGroup('FEATURED_VIDEO_VIEWABILITY_VARIANTS', 'RENDER_ORDER')) {
-			scheduleOnce('afterRender', () => {
-				updateFeaturedVideoPosition();
-			});
-		}
-	}),
+	/**
+	 * @returns {boolean}
+	 */
+	isUserLangEn: equal('currentUser.language', 'en'),
+	shouldShowFandomAppSmartBanner: and('isUserLangEn', 'wikiVariables.enableFandomAppSmartBanner'),
+	isFandomAppSmartBannerVisible: and('shouldShowFandomAppSmartBanner', 'smartBannerVisible'),
 
 	/**
 	 * @returns {void}
