+++ app/components/article-featured-video-jwplayer.js
@@ -1,0 +1,93 @@
+import {inject as service} from '@ember/service';
+import Component from '@ember/component';
+import {on} from '@ember/object/evented';
+import {observer, computed} from '@ember/object';
+import {htmlSafe} from '@ember/string';
+import VideoLoader from '../modules/video-loader';
+import extend from '../utils/extend';
+import config from '../config/environment';
+
+export default Component.extend({
+	ads: service(),
+	wikiVariables: service(),
+
+	autoplayCookieName: 'featuredVideoAutoplay',
+	captionsCookieName: 'featuredVideoCaptions',
+	playerCookieExpireDays: 14,
+	placeholderImage: computed('model', function () {
+		return this.get('model.embed.jsParams.playlist.0.image');
+	}),
+	placeholderStyle: computed('placeholderImage', function () {
+		return htmlSafe(`background-image: url(${this.get('placeholderImage')})`);
+	}),
+
+	// when navigating from one article to another with video, we need to destroy player and
+	// reinitialize it as component itself is not destroyed. Could be done with didUpdateAttrs
+	// hook, however it is fired twice with new attributes.
+	videoIdObserver: on('didInsertElement', observer('model.embed.jsParams.videoId', function () {
+		this.destroyVideoPlayer();
+		this.initVideoPlayer();
+	})),
+
+	init() {
+		this._super(...arguments);
+
+		this.set('videoContainerId', `jwplayer-article-video-${new Date().getTime()}`);
+	},
+
+	/**
+	 * @param {Object} player
+	 * @returns {void}
+	 */
+	onCreate(player) {
+		this.player = player;
+
+		this.player.on('autoplayToggle', ({enabled}) => {
+			this.setCookie(this.get('autoplayCookieName'), (enabled ? '1' : '0'));
+		});
+
+		this.player.on('captionsSelected', ({selectedLang}) => {
+			this.setCookie(this.get('captionsCookieName'), selectedLang);
+		});
+	},
+
+	/**
+	 * @returns {void}
+	 */
+	initVideoPlayer() {
+		const model = this.get('model.embed'),
+			jsParams = {
+				autoplay: $.cookie(this.get('autoplayCookieName')) !== '0',
+				selectedCaptionsLanguage: $.cookie(this.get('captionsCookieName')),
+				adTrackingParams: {
+					adProduct: this.get('ads.noAds') ? 'featured-video-no-preroll' : 'featured-video-preroll',
+					slotName: 'FEATURED'
+				},
+				containerId: this.get('videoContainerId'),
+				noAds: this.get('ads.noAds'),
+				onCreate: this.onCreate.bind(this),
+				lang: this.get('wikiVariables.language.content')
+			},
+			data = extend({}, model, {jsParams}),
+			videoLoader = new VideoLoader(data);
+
+		videoLoader.loadPlayerClass();
+	},
+
+	/**
+	 * @returns {void}
+	 */
+	destroyVideoPlayer() {
+		if (this.player) {
+			this.player.remove();
+		}
+	},
+
+	setCookie(cookieName, cookieValue) {
+		$.cookie(cookieName, cookieValue, {
+			expires: this.get('playerCookieExpireDays'),
+			path: '/',
+			domain: config.cookieDomain
+		});
+	}
+});
