+++ app/components/portable-infobox.js
@@ -3,7 +3,6 @@ import Component from '@ember/component';
 import ViewportMixin from '../mixins/viewport';
 import RenderComponentMixin from '../mixins/render-component';
 import {track, trackActions} from '../utils/track';
-import {inGroup} from '../modules/abtest';
 
 export default Component.extend(
 	RenderComponentMixin,
@@ -45,18 +44,9 @@ export default Component.extend(
 		collapsedHeight: computed('viewportDimensions.width', 'viewportDimensions.height', function () {
 			const deviceWidth = this.get('viewportDimensions.width'),
 				deviceHeight = this.get('viewportDimensions.height'),
-				isLandscape = deviceWidth > deviceHeight,
-				calculatedHeight = Math.floor((isLandscape ? deviceHeight : deviceWidth) * 16 / 9) + 100,
-				maximumRequiredHeight = 500;
+				isLandscape = deviceWidth > deviceHeight;
 
-			/**
-			 * FIXME FEATURED VIDEO A/B TEST ONLY
-			 */
-			if (inGroup('FEATURED_VIDEO_VIEWABILITY_VARIANTS', 'PAGE_PLACEMENT')) {
-				return calculatedHeight > maximumRequiredHeight ? maximumRequiredHeight : calculatedHeight;
-			}
-
-			return calculatedHeight;
+			return Math.floor((isLandscape ? deviceHeight : deviceWidth) * 16 / 9) + 100;
 		}),
 
 		didInsertElement() {
