+++ app/components/site-head.js
@@ -18,11 +18,6 @@ export default Component.extend(
 
 		ads: service(),
 		notifications: service(),
-		smartBanner: service(),
-
-		smartBannerVisible: readOnly('smartBanner.smartBannerVisible'),
-		shouldShowFandomAppSmartBanner: readOnly('smartBanner.shouldShowFandomAppSmartBanner'),
-		isFandomAppSmartBannerVisible: readOnly('smartBanner.isFandomAppSmartBannerVisible'),
 
 		headroomOptions: {
 			classes: {
