+++ app/index.html
@@ -7,10 +7,20 @@
 {{content-for 'head-footer'}}
 </head>
 <body class="mobile-wiki">
-<div id="pre-featured-video-wrapper"><div id="pre-featured-video"></div></div>
 <div id="ember-container">{{content-for 'body'}}</div>
 
 {{content-for 'measure-first-render'}}
+
+<script>
+	(function (M) {
+		if (typeof FastBoot === 'undefined' && M.getFromShoebox('serverError')) {
+			// No need to load Ember in browser on server error page
+			return;
+		}
+		M.loadScript('/mobile-wiki/assets/vendor.js', false, false, 'anonymous');
+		M.loadScript('/mobile-wiki/assets/mobile-wiki.js', false, false, 'anonymous');
+	})(window.M);
+</script>
 
 {{content-for 'fastboot-inline-scripts-load-svg'}}
 <script>
@@ -21,34 +31,7 @@
 </script>
 
 <script src="/__load/-/only=scripts/wikia.ext.instantGlobals,instantGlobalsOverride,abtesting,abtest"></script>
-<script>
-	(function (M) {
-		if (window.Wikia.AbTest.inGroup('FEATURED_VIDEO_VIEWABILITY_VARIANTS', 'RENDER_ORDER') || (typeof FastBoot === 'undefined' && M.getFromShoebox('serverError'))) {
-			// No need to load Ember in browser on server error page
-			return;
-		}
-		M.loadScript('/mobile-wiki/assets/vendor.js', false, false, 'anonymous');
-		M.loadScript('/mobile-wiki/assets/mobile-wiki.js', false, false, 'anonymous');
-	})(window.M);
-</script>
 
-<script>
-	/**
-	 * FIXME FEATURED VIDEO A/B TEST ONLY
-	 *
-	 * This is the earliest moment when we can get AB group
-	 * Normally we prerender video placeholder, but in this specific variant of AB Testing video is
-	 * placed differently so we have to hide the prerendered placeholder.
-	 */
-	if (window.Wikia.AbTest.inGroup('FEATURED_VIDEO_VIEWABILITY_VARIANTS', 'PAGE_PLACEMENT')) {
-		document.querySelector('.article-featured-video').style.display = 'none';
-	}
-
-	if (window.Wikia.AbTest.inGroup('FEATURED_VIDEO_VIEWABILITY_VARIANTS', 'RENDER_ORDER')) {
-		M.loadScript('/mobile-wiki/assets/abtest/featured-video-render-order.js', false, false,
-			'anonymous');
-	}
-</script>
 {{content-for 'fastboot-inline-scripts-body-bottom'}}
 {{content-for 'tracking-ua'}}
 {{content-for 'tracking-internal'}}
