--- app/modules/abtest/offset.js
@@ -1,21 +1,0 @@
-export default function offset(element) {
-	if (!element) {
-		return;
-	}
-
-	// Return zeros for disconnected and hidden (display: none) elements (gh-2310)
-	// Support: IE <=11 only
-	// Running getBoundingClientRect on a
-	// disconnected node in IE throws an error
-	if (!element.getClientRects().length) {
-		return {top: 0, left: 0};
-	}
-
-	// Get document-relative position by adding viewport scroll to viewport-relative gBCR
-	const rect = element.getBoundingClientRect(),
-		win = element.ownerDocument.defaultView;
-	return {
-		top: rect.top + win.pageYOffset,
-		left: rect.left + win.pageXOffset
-	};
-}
