+++ app/styles/component/index.scss
@@ -36,7 +36,7 @@
 @import 'widget-discussion';
 @import 'wikia-user-profile';
 @import 'show-error';
-@import 'article-featured-video';
+@import 'article-featured-video-jwplayer';
 @import 'recirculation';
 @import 'fandom-app-smart-banner';
 @import 'jwplayer-in-article-video';
