+++ app/templates/application.hbs
@@ -19,8 +19,10 @@
 	loadRandomArticle=(action 'loadRandomArticle')
 	setQueryParam=(action 'setQueryParam')
 	siteHeadNoShadow=siteHeadNoShadow
+	smartBannerVisible=smartBannerVisible
 	themeBar=themeBar
 	toggleDrawer=(action 'toggleDrawer')
+	toggleSmartBanner=(action 'toggleSmartBanner')
 	toggleSiteHeadShadow=(action 'toggleSiteHeadShadow')
 	toggleUserMenu=(action 'toggleUserMenu')
 	userMenuVisible=userMenuVisible
