+++ app/templates/components/article-featured-video-jwplayer.hbs
@@ -1,0 +1,1 @@
+<div class="article-featured-video-jwplayer" id="{{videoContainerId}}" style={{placeholderStyle}}></div>
