--- app/templates/components/article-featured-video.hbs
@@ -1,9 +1,0 @@
-<div class="article-featured-video__on-scroll-video-wrapper">
-	<div class="article-featured-video__video-content" style={{placeholderStyle}} id="{{videoContainerId}}"></div>
-	<div class="article-featured-video__close-button" {{action 'dismissPlayer'}}>
-		<svg class="wds-icon wds-icon-small">
-			<use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="#wds-icons-cross"></use>
-		</svg>
-	</div>
-</div>
-<div class="article-featured-video__on-scroll-placeholder"></div>
