+++ app/templates/components/site-head.hbs
@@ -1,5 +1,6 @@
 {{#if shouldShowFandomAppSmartBanner}}
 	{{fandom-app-smart-banner
+		toggleVisibility=toggleSmartBannerVisibility
 		isVisible=smartBannerVisible
 	}}
 {{/if}}
