+++ app/utils/track.js
@@ -137,7 +137,7 @@ export function track(params) {
 	if (!window.location) {
 		return;
 	}
-	params = Object.assign({
+	params = $.extend({
 		ga_action: action,
 		ga_category: category,
 		ga_label: label,
@@ -157,7 +157,7 @@ export function track(params) {
 	}
 
 	if (trackingMethod === 'both' || trackingMethod === 'internal') {
-		params = Object.assign({}, context, params);
+		params = $.extend({}, context, params);
 		M.tracker.Internal.track(isPageView(category) ? 'view' : 'special/trackingevent', params);
 	}
 }
